made the return type as double for fine/fee getters and deleted the redundant ApplicationUser file

+++ code/framework/CheckinCheckoutFramework/src/main/java/cs525/project/fujframework/core/Product.java
@@ -52,9 +52,9 @@ public interface Product {
 	/**
 	 * returns the rental fee per day
 	 * 
-	 * @return fee per day as integer
+	 * @return fee per day as double
 	 */
-	public int getRentalFeePerDay();
+	public double getRentalFeePerDay();
 
 	/**
 	 * sets the due fine per day
@@ -66,8 +66,8 @@ public interface Product {
 	/**
 	 * returns the fine rental per day
 	 * 
-	 * @return fine per day as integer
+	 * @return fine per day as double
 	 */
-	public int getOverDueFinePerDay();
+	public double getOverDueFinePerDay();
 
 }
